pyproject.toml
[tool.poetry]
name = "cookie"
version = "0.1.0"
description = ""
authors = ["Robin Kahlow <robin@ntropy.com>"]
readme = "README.md"
packages = [{include = "cookie"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
ntropy-sdk = "^4.18.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.9"
discord-py = "^2.2.2"
tabulate = "^0.9.0"
more-itertools = "^9.1.0"
psycopg2-binary = "^2.8"
cachetools = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
